+++ sound-mind-studio/tests/main.cpp
@@ -27,6 +27,7 @@
 #include "test_gradient_editor_widget.h"
 #include "test_grid_config.h"
 #include "test_grid_panel.h"
+#include "test_history_panel.h"
 #include "test_image_scale_picker_dialog.h"
 #include "test_import_export.h"
 #include "test_import_helpers.h"
@@ -160,6 +161,9 @@ int main(int argc, char** argv) {
 
     GridPanelTest gridPanelTest;
     status |= QTest::qExec(&gridPanelTest, argc, argv);
+
+    HistoryPanelTest historyPanelTest;
+    status |= QTest::qExec(&historyPanelTest, argc, argv);
 
     GridConfigTest gridConfigTest;
     status |= QTest::qExec(&gridConfigTest, argc, argv);