+++ sound-mind-studio/tests/test_history_panel.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <QObject>
+
+class HistoryPanelTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void freshPanelShowsOnlyTheStartRow();
+    void setHistoryShowsEachDescriptionAndHighlightsTheCurrentIndex();
+    void setHistoryUsesAPlaceholderForAnEmptyDescription();
+    void doubleClickingARowEmitsJumpRequested();
+};