+++ sound-mind-studio/tests/test_undo_stack.h
@@ -15,4 +15,6 @@ private slots:
     void redoWhenNothingUndoneIsANoOp();
     void pushingAfterAnUndoDiscardsTheRedoTail();
     void clearDiscardsEveryCommandAndResetsAvailability();
+    void countCurrentIndexAndDescriptionAtReflectPushedCommands();
+    void jumpToMovesDirectlyInvokingEveryInBetweenCallback();
 };